fix: Deep-copy tensors when writing to `state_dict`
This fixes a bug for quantized models. When `state_dict` stored the tensors
as references, it saved quantized weights as having been post-processed
(notably `q_scale_max` being divided by 256 in `ext.make_q_matrix()`).
This caused it to be post-processed an additional second time when
deserialized, causing underflow for `q_scale_max`.